Move HostnameValidator to ghinstance
