+++ internal/ghinstance/host_test.go
@@ -2,6 +2,8 @@ package ghinstance
 
 import (
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestOverridableDefault(t *testing.T) {
@@ -97,6 +99,50 @@ func TestNormalizeHostname(t *testing.T) {
 	}
 }
 
+func TestHostnameValidator(t *testing.T) {
+	tests := []struct {
+		name     string
+		input    interface{}
+		wantsErr bool
+	}{
+		{
+			name:     "valid hostname",
+			input:    "internal.instance",
+			wantsErr: false,
+		},
+		{
+			name:     "hostname with slashes",
+			input:    "//internal.instance",
+			wantsErr: true,
+		},
+		{
+			name:     "empty hostname",
+			input:    "   ",
+			wantsErr: true,
+		},
+		{
+			name:     "hostname with colon",
+			input:    "internal.instance:2205",
+			wantsErr: true,
+		},
+		{
+			name:     "non-string hostname",
+			input:    62,
+			wantsErr: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := HostnameValidator(tt.input)
+			if tt.wantsErr {
+				assert.Error(t, err)
+				return
+			}
+			assert.Equal(t, nil, err)
+		})
+	}
+}
 func TestGraphQLEndpoint(t *testing.T) {
 	tests := []struct {
 		host string
